Split data_science pipeline into training and evaluation
