+++ demo/delay-prediction/conf/base/parameters_training.yml
@@ -1,0 +1,5 @@
+# This is a boilerplate parameters config generated for pipeline 'training'
+# using Kedro 0.19.6.
+#
+# Documentation for this file format can be found in "Parameters"
+# Link: https://docs.kedro.org/en/0.19.6/configuration/parameters.html
