+++ demo/delay-prediction/src/delay_prediction/pipelines/evaluation/__init__.py
@@ -1,0 +1,5 @@
+from .pipeline import create_pipeline
+
+__all__ = ["create_pipeline"]
+
+__version__ = "0.1"
