+++ demo/delay-prediction/src/delay_prediction/pipelines/evaluation/pipeline.py
@@ -1,0 +1,13 @@
+from kedro.pipeline import Pipeline, node, pipeline
+
+from .nodes import evaluate_model
+
+
+def create_pipeline(**kwargs) -> Pipeline:
+    return pipeline([
+        node(
+            func=evaluate_model,
+            inputs=["pipe", "X_test", "y_test"],
+            outputs=None,
+        ),
+    ])
