+++ demo/delay-prediction/src/delay_prediction/pipelines/training/pipeline.py
@@ -1,6 +1,6 @@
 from kedro.pipeline import Pipeline, node, pipeline
 
-from .nodes import evaluate_model, split_data, train_model
+from .nodes import split_data, train_model
 
 
 def create_pipeline(**kwargs) -> Pipeline:
@@ -15,11 +15,6 @@ def create_pipeline(**kwargs) -> Pipeline:
                 func=train_model,
                 inputs=["X_train", "y_train"],
                 outputs="pipe",
-            ),
-            node(
-                func=evaluate_model,
-                inputs=["pipe", "X_test", "y_test"],
-                outputs=None,
             ),
         ]
     )
